+++ src/api/views/user.py
@@ -3,30 +3,17 @@
 Created on 23/01/2024 at 17:53:44(+00:00).
 """
 
-import logging
-from datetime import timedelta
-
-from codeforlife.mail import send_mail
-from codeforlife.permissions import (
-    OR,
-    AllowAny,
-    AllowNone,
-    IsCronRequestFromGoogle,
-)
+from codeforlife.permissions import OR, AllowAny, AllowNone
 from codeforlife.request import Request
 from codeforlife.response import Response
 from codeforlife.user.models import User
 from codeforlife.user.permissions import IsIndependent, IsTeacher
 from codeforlife.user.views import UserViewSet as _UserViewSet
-from codeforlife.views import action, cron_job
+from codeforlife.views import action
 from django.conf import settings
-from django.db.models import F, Q
-from django.urls import reverse
-from django.utils import timezone
 from rest_framework import status
 from rest_framework.serializers import ValidationError
 
-from ..auth import email_verification_token_generator
 from ..serializers import (
     CreateUserSerializer,
     HandleIndependentUserJoinClassRequestSerializer,
@@ -62,15 +49,6 @@ class UserViewSet(_UserViewSet):
             and "requesting_to_join_class" in self.request.data
         ):
             return [IsIndependent()]
-        if self.action in [
-            "send_1st_verify_email_reminder",
-            "send_2nd_verify_email_reminder",
-            "anonymize_unverified_accounts",
-            "send_1st_inactivity_reminder",
-            "send_2nd_inactivity_reminder",
-            "send_final_inactivity_reminder",
-        ]:
-            return [IsCronRequestFromGoogle()]
 
         return super().get_permissions()
 
@@ -208,224 +186,3 @@ class UserViewSet(_UserViewSet):
     handle_join_class_request = _UserViewSet.update_action(
         "handle_join_class_request"
     )
-
-    def _get_unverified_users(self, days: int, same_day: bool):
-        now = timezone.now()
-
-        # All expired unverified users.
-        user_queryset = User.objects.filter(
-            date_joined__lte=now - timedelta(days=days),
-            userprofile__is_verified=False,
-        )
-        if same_day:
-            user_queryset = user_queryset.filter(
-                date_joined__gt=now - timedelta(days=days + 1)
-            )
-
-        teacher_queryset = user_queryset.filter(
-            new_teacher__isnull=False,
-            new_student__isnull=True,
-        )
-        independent_student_queryset = user_queryset.filter(
-            new_teacher__isnull=True,
-            new_student__class_field__isnull=True,
-        )
-
-        return teacher_queryset, independent_student_queryset
-
-    def _send_verify_email_reminder(self, days: int, campaign_name: str):
-        teacher_queryset, indy_queryset = self._get_unverified_users(
-            days, same_day=True
-        )
-
-        user_queryset = teacher_queryset.union(indy_queryset)
-        user_count = user_queryset.count()
-
-        logging.info("%d emails unverified.", user_count)
-
-        if user_count > 0:
-            sent_email_count = 0
-            for user_fields in user_queryset.values("id", "email").iterator(
-                chunk_size=500
-            ):
-                url = settings.SERVICE_BASE_URL + reverse(
-                    "user-verify-email-address",
-                    kwargs={
-                        "pk": user_fields["id"],
-                        "token": email_verification_token_generator.make_token(
-                            user_fields["id"], user_fields["email"]
-                        ),
-                    },
-                )
-
-                try:
-                    send_mail(
-                        campaign_id=settings.DOTDIGITAL_CAMPAIGN_IDS[
-                            campaign_name
-                        ],
-                        to_addresses=[user_fields["email"]],
-                        personalization_values={"VERIFICATION_LINK": url},
-                    )
-
-                    sent_email_count += 1
-                # pylint: disable-next=broad-exception-caught
-                except Exception as ex:
-                    logging.exception(ex)
-
-            logging.info("Sent %d/%d emails.", sent_email_count, user_count)
-
-        return Response()
-
-    @cron_job
-    def send_1st_verify_email_reminder(self, request: Request):
-        """
-        Send the first reminder email to all users who have not verified their
-        email address.
-        """
-        return self._send_verify_email_reminder(
-            days=7, campaign_name="Verify new user email - first reminder"
-        )
-
-    @cron_job
-    def send_2nd_verify_email_reminder(self, request: Request):
-        """
-        Send the second reminder email to all users who have not verified their
-        email address.
-        """
-        return self._send_verify_email_reminder(
-            days=14, campaign_name="Verify new user email - second reminder"
-        )
-
-    @cron_job
-    def anonymize_unverified_accounts(self, request: Request):
-        """Anonymize all users who have not verified their email address."""
-        user_queryset = User.objects.filter(is_active=True)
-        user_count = user_queryset.count()
-
-        teacher_queryset, indy_queryset = self._get_unverified_users(
-            days=int(request.query_params.get("days", 19)),
-            same_day=False,
-        )
-        teacher_count = teacher_queryset.count()
-        indy_count = indy_queryset.count()
-
-        for user in teacher_queryset.union(indy_queryset).iterator(
-            chunk_size=100
-        ):
-            try:
-                user.anonymize()
-            # pylint: disable-next=broad-exception-caught
-            except Exception as ex:
-                logging.error("Failed to anonymise user with id: %d", user.id)
-                logging.exception(ex)
-
-        logging.info(
-            "%d unverified users anonymised.",
-            user_count - user_queryset.count(),
-        )
-
-        # Use data warehouse in new system.
-        # pylint: disable-next=import-outside-toplevel
-        from common.models import (  # type: ignore[import-untyped]
-            DailyActivity,
-            TotalActivity,
-        )
-
-        activity_today = DailyActivity.objects.get_or_create(
-            date=timezone.now().date()
-        )[0]
-        activity_today.anonymised_unverified_teachers = teacher_count
-        activity_today.anonymised_unverified_independents = indy_count
-        activity_today.save()
-        TotalActivity.objects.update(
-            anonymised_unverified_teachers=F("anonymised_unverified_teachers")
-            + teacher_count,
-            anonymised_unverified_independents=F(
-                "anonymised_unverified_independents"
-            )
-            + indy_count,
-        )
-
-        return Response()
-
-    def _get_inactive_users(self, days: int):
-        now = timezone.now()
-
-        # All users who haven't logged in in X days OR who've never logged in
-        # and registered over X days ago.
-        user_queryset = User.objects.filter(
-            Q(
-                last_login__isnull=False,
-                last_login__lte=now - timedelta(days=days),
-                last_login__gt=now - timedelta(days=days + 1),
-            )
-            | Q(
-                last_login__isnull=True,
-                date_joined__lte=now - timedelta(days=days),
-                date_joined__gt=now - timedelta(days=days + 1),
-            )
-        )
-
-        return user_queryset.exclude(email__isnull=True).exclude(email="")
-
-    def _send_inactivity_reminder(self, days: int, campaign_name: str):
-        user_queryset = self._get_inactive_users(days)
-        user_count = user_queryset.count()
-
-        logging.info("%d inactive users after %d days.", user_count, days)
-
-        if user_count > 0:
-            sent_email_count = 0
-            for email in user_queryset.values_list("email", flat=True).iterator(
-                chunk_size=500
-            ):
-                try:
-                    send_mail(
-                        campaign_id=settings.DOTDIGITAL_CAMPAIGN_IDS[
-                            campaign_name
-                        ],
-                        to_addresses=[email],
-                    )
-
-                    sent_email_count += 1
-                # pylint: disable-next=broad-exception-caught
-                except Exception as ex:
-                    logging.exception(ex)
-
-            logging.info(
-                "Reminded %d/%d inactive users.", sent_email_count, user_count
-            )
-
-        return Response()
-
-    @cron_job
-    def send_1st_inactivity_reminder(self, request: Request):
-        """
-        Send the first reminder email to teachers and independent users who
-        haven't been active in a while.
-        """
-        return self._send_inactivity_reminder(
-            days=730, campaign_name="Inactive users on website - first reminder"
-        )
-
-    @cron_job
-    def send_2nd_inactivity_reminder(self, request: Request):
-        """
-        Send the second reminder email to teachers and independent users who
-        haven't been active in a while.
-        """
-        return self._send_inactivity_reminder(
-            days=973,
-            campaign_name="Inactive users on website - second reminder",
-        )
-
-    @cron_job
-    def send_final_inactivity_reminder(self, request: Request):
-        """
-        Send the final reminder email to teachers and independent users who
-        haven't been active in a while.
-        """
-        return self._send_inactivity_reminder(
-            days=1065,
-            campaign_name="Inactive users on website - final reminder",
-        )
