+++ src/sso/urls.py
@@ -5,7 +5,7 @@ Created on 01/12/2023 at 16:02:53(+00:00).
 
 from django.urls import include, path, re_path
 
-from .views import ClearExpiredView, LoginView
+from .views import LoginView
 
 urlpatterns = [
     path(
@@ -17,11 +17,6 @@ urlpatterns = [
                     r"^(?P<form>login-with-email|login-with-otp|login-with-otp-bypass-token|login-as-student|auto-login-as-student)/$",
                     LoginView.as_view(),
                     name="session-login",
-                ),
-                path(
-                    "clear-expired/",
-                    ClearExpiredView.as_view(),
-                    name="session-clear-expired",
                 ),
             ]
         ),
